# WebApiTransJ/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "WebApiTransJ, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Unidades/RegistrarUnidad": {
      "post": {
        "tags": [
          "Unidades"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UnidadEntity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Unidades/ActualizarUnidad": {
      "put": {
        "tags": [
          "Unidades"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UnidadEntity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Unidades/CambiarEstadoUnidad": {
      "put": {
        "tags": [
          "Unidades"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "IdUnidad",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "IdUsuario",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Unidades/listarUnidades": {
      "get": {
        "tags": [
          "Unidades"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "IdSocio",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Unidades/listarUnidadesXId": {
      "get": {
        "tags": [
          "Unidades"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "IdUnidad",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Unidades/ListaViajesUnidad": {
      "get": {
        "tags": [
          "Unidades"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "IdUnidad",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "tipMes",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuario/RegistrarUsuario": {
      "post": {
        "tags": [
          "Usuario"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UsuarioEntity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuario/ActualizarUsuario": {
      "put": {
        "tags": [
          "Usuario"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UsuarioEntity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Usuario/CambiarEstadoUsuario": {
      "put": {
        "tags": [
          "Usuario"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "IdUsuario",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "UsuarioModificacion",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuario/ListarUsuariosXId": {
      "get": {
        "tags": [
          "Usuario"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "IdUsuario",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "UnidadEntity": {
      "type": "object",
      "properties": {
        "pCodigoUnidad": {
          "type": "string"
        },
        "pColorUnidad": {
          "type": "string"
        },
        "pNoPlaca": {
          "type": "string"
        },
        "pIdUsuario": {
          "type": "string"
        },
        "pIdUnidad": {
          "format": "int32",
          "type": "integer"
        },
        "pIdSocio": {
          "format": "int32",
          "type": "integer"
        },
        "pModelo": {
          "type": "string"
        },
        "pIdMarca": {
          "format": "int32",
          "type": "integer"
        },
        "pTransaccionEstado": {
          "format": "int32",
          "type": "integer"
        },
        "pTransaccionMensaje": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "UsuarioEntity": {
      "type": "object",
      "properties": {
        "pIdUsuario": {
          "type": "string"
        },
        "pPNombre": {
          "type": "string"
        },
        "sNombre": {
          "type": "string"
        },
        "pPApellido": {
          "type": "string"
        },
        "pSApellido": {
          "type": "string"
        },
        "pDireccion": {
          "type": "string"
        },
        "pNoDPI": {
          "type": "string"
        },
        "pNoTelefono": {
          "type": "string"
        },
        "pCorreo": {
          "type": "string"
        },
        "pContrasenia": {
          "type": "string"
        },
        "pIdRol": {
          "format": "int32",
          "type": "integer"
        },
        "pIdOrganizacion": {
          "format": "int32",
          "type": "integer"
        },
        "pIdUsuarioCreacion": {
          "type": "string"
        },
        "pTransaccionEstado": {
          "format": "int32",
          "type": "integer"
        },
        "pTransaccionMensaje": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
  }
}